Initial Logic BugFix on ExpenseAccount Creation/Matching:
When the AI receives an description of an purchase. The current prompt
specifies to create an account, but not how. So it could create an
account that is entirely different from the description. in those cases,
the `autocomplete/accounts` it's not going to work because the query
from the description vs the expense account is not the same.

We are not passing the Existing Expense Accounts to the AI model for
avoiding bloating the context window and avoiding the possibility for
wrong picks at all costs.